WIP: xilinx: zynqmp: Add support for K26 and KV260
Add support for K26 and KV260
The support is split to 4 files for the following reason.

-k26-revA.dts - This is base DTS file which describes only SOM with
few exceptions.
a) GPIOs
b) uart at MIO34/MIO35
c) Aliases are filled even by IPs which are not enabled to have consistent
numbering scheme

-k26-revA-u-boot.dts - DTS file which targets only U-Boot project and
 which is automatically pick up by u-boot at build time. It enables SD card
 which is available on CC and it disable all HS modes capabilities to make
 sure that u-boot is capable to read data from SD without a need to setup
 tap delays for custom CCs.

-kv260-revA.dts - it is device tree overlay automatically built as dtbo
 which describes CC and all features on it.

-kv260-revA-pl.dts - ML Vision description for hardware i2c switch which is
 connected to PL. As of today it is here just for reference that there is a
 node which needs to be described and handled by SW.

Pincntrl description is not present but it could come in future especially
in connection to SD card connection.

There is also high chance that there is going to be a need to detect CC
from u-boot that's why u-boot DTS could also contain eeprom available at CC
for detection. It doesn't need to be there but it is good to list it not to
go around driver model in u-boot.

For checking DTBo fdtoverlay command can be used to make sure that base DT
and overlays are compatible to each other:
fdtoverlay -v -i arch/arm/dts/zynqmp-sm-k26-revA.dtb \
arch/arm/dts/zynqmp-sk-kv260-revA.dtbo -o /tmp/1.dtb

For building u-boots dtb file just run make.
For building Linux dtb files don't include u-boot.dtsi in build

Signed-off-by: Michal Simek <michal.simek@xilinx.com>
---
Note: patch is labeled as WIP and will don't go to master till we test it
on revA boards. Patch is going to be added only to master-next-test for
reference when reviewed.

This patch also contain -pl dts file which is just for reference. Let's see
how we will handle it.

http://jira.xilinx.com/browse/NSPR-1829

v2:
- rename them to sk-kv260 and sm-k26 - is internally called Vision SOM and
  ML Starter kit
- SOM wire CC eeprom

v3:
- Another rename to sck-kv-g and sm-k26 with proper revision format

v4:
- Get rid of uart0 setting. Default uart is uart1. This has been changed in
  20200727 mlcc CC
- Also update overlay - remove uart1 enabling and s/fragment@/fragment/g

v5:
- Lower frequency for QSPI. MIO6(loopback pin) is not assigned to QSPI
  controller that's why maximum frequncy is 40MHz

v6:
- Fix spi IP version it is spi1 instead of spi0

v7:
- eth phy at 5 instead of 1

v8:
- add missing compatible string

v9:
- update clk input lines for GTs