arm64: zynqmp: Add support for SVD devices
SVD (Software Virtual Devices) are using different name which can't be
handled via zynqmp_devices structure. That's why introduce
zynqmp_detect_svd_name() which checks ID code for these devices and show
proper name for them.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>
---
Based on Alveo
https://confluence.xilinx.com/display/XIP/Alveo+Device+Addition
and
https://confluence.xilinx.com/display/XPTM/SOM+Silicon+Obfuscation

I have added IDs also for Alveo cards but we don't need to push it.
sck24 is also added as zu3 + 100 SVD but we also don't need to push it.